Most Current Build
Before running make sure to set the path of the gun scripts. To get the file's path press (shift +right click) then click "Copy as Path". Paste the new path in place of the old one which should be right under the corresponding labels for the guns.
You may have issues with the crosshair not showing up. If so move "Cross.png" image to your desktop.